fix: safely load assemblies
Change assembly.GetTypes() to a more lenient approach "GetLoadableTypes()" where ReflectionTypeLoadException is being ignored, and we load only the types that we are able to load.

The reason for this is that there may be some dlls that will fail to load which will fail the entire flow.
For example the following error (which happened):
ReflectionTypeLoadException : Unable to load one or more of the requested types.\nCould not load type 'SqlGuidCaster' from assembly 'System.Data.SqlClient, Version=4.6.1.6, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a' because it contains an object field at offset 0 that is incorrectly aligned or overlapped by a non-object field."

Signed-off-by: Yuval Bavli <Yuval.Bavli@docusign.com>